Check for existing sources before copying
It should be possible to use the vlogmodulefiles attribute to specify
sources that are externally generated, e.g. to support blackboxed
verilog modules included in a chisel design.

+++ rtl/__init__.py
@@ -607,8 +607,11 @@ class rtl(thesdk,metaclass=abc.ABCMeta):
             for modfile in self.vlogmodulefiles:
                 srcfile = os.path.join(self.vlogsrcpath, modfile)
                 dstfile = os.path.join(self.rtlsimpath, modfile)
-                self.print_log(type='I', msg='Copying %s to %s' % (srcfile, dstfile))
-                shutil.copyfile(srcfile, dstfile)
+                if os.path.isfile(dstfile):
+                    self.print_log(type='I', msg='Using externally generated source: %s' % modfile)
+                else:
+                    self.print_log(type='I', msg='Copying %s to %s' % (srcfile, dstfile))
+                    shutil.copyfile(srcfile, dstfile)
 
         # nothing generates vhdl so simply copy all files to rtlsimpath
         elif self.model == 'vhdl':
@@ -616,8 +619,11 @@ class rtl(thesdk,metaclass=abc.ABCMeta):
             for entfile in self.vhdlentityfiles:
                 srcfile = os.path.join(self.vlogsrcpath, entfile)
                 dstfile = os.path.join(self.rtlsimpath, entfile)
-                self.print_log(type='I', msg='Copying %s to %s' % (srcfile, dstfile))
-                shutil.copyfile(srcfile, dstfile)
+                if os.path.isfile(dstfile):
+                    self.print_log(type='I', msg='Using externally generated source: %s' % entfile)
+                else:
+                    self.print_log(type='I', msg='Copying %s to %s' % (srcfile, dstfile))
+                    shutil.copyfile(srcfile, dstfile)
 
         # flush cached writes to disk
         output = subprocess.check_output("sync %s" % self.rtlsimpath, shell=True)
